eq by range index
