+++ rust/poker_eval/src/core/hole_cards.rs
@@ -11,6 +11,8 @@ pub struct HoleCards {
     //card_lo: Card
 }
 
+pub const SIMPLE_RANGE_INDEX_LEN: usize = 169;
+
 impl HoleCards {
     pub fn new(card1: Card, card2: Card) -> Result<Self, PokerError> {
         let card1_index: u8 = card1.into();
