+++ rust/poker_eval/src/web/analyzer.rs
@@ -126,19 +126,20 @@ impl flop_analyzer {
         Ok(())
     }
 
-    pub fn set_player_range(&mut self, player_idx: usize, range_str: &str) {
+    pub fn set_player_range(&mut self, player_idx: usize, range_str: &str) -> Result<(), PokerError> {
         info!("set_player_range: {} [{}]", player_idx, range_str);
 
         if range_str.is_empty() {
-            warn!("set_player_range: empty string");
-            return;
+            return Err(PokerError::from_str("set_player_range: empty string"));
         }
-        let range_set = range_string_to_set(range_str);
+        let range_set = range_string_to_set(range_str)?;
 
         info!("% is {}", range_set.count_ones() as f64 / 2652.0);
 
         self.player_info[player_idx].range_set = range_set;
         self.player_info[player_idx].range_string = range_str.to_string();
+
+        Ok(())
     }
 
     pub fn set_player_state(&mut self, player_idx: usize, state: u8) {
@@ -414,310 +415,5 @@ impl flop_analyzer {
         }
 
         Ok(num_chosen)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::{card_u8s_from_string, web::analyzer::PlayerPreFlopState, Rank};
-
-    fn assert_equity(equity: f64, target: f64, tolerance: f64) {
-        let passed = (equity - target).abs() < tolerance;
-        if !passed {
-            println!("assert_equity failed: {} != {}", equity, target);
-        }
-        assert!(passed);
-    }
-
-    #[test]
-    fn test_heads_up_both_with_hole_cards() {
-        let mut analyzer = super::flop_analyzer::new();
-        analyzer.reset();
-        let tolerance = 0.1;
-
-        analyzer.set_player_state(0, PlayerPreFlopState::UseHoleCards as u8);
-        analyzer.set_player_state(3, PlayerPreFlopState::UseHoleCards as u8);
-
-        analyzer
-            .set_player_cards(0, card_u8s_from_string("7h 6s").as_slice())
-            .unwrap();
-
-        analyzer
-            .set_player_cards(3, card_u8s_from_string("Th 9h").as_slice())
-            .unwrap();
-
-        analyzer
-            .set_board_cards(card_u8s_from_string("Qs Ts 7c").as_slice())
-            .unwrap();
-
-        let num_it = 10_000;
-        let f_results = analyzer.build_results();
-        let f_results = analyzer.simulate_flop(num_it, f_results).unwrap();
-
-        let results = &f_results.flop_results;
-
-        // let results = analyzer
-        //     .player_info
-        //     .iter()
-        //     .map(|p| &p.results)
-        //     .collect_vec();
-
-        assert_eq!(results[0].street_rank_results[2].num_iterations, num_it);
-        assert_eq!(0, results[0].player_index);
-        assert_equity(
-            100.0 * results[0].street_rank_results[2].win_eq / num_it as f64,
-            21.92,
-            tolerance,
-        );
-
-        assert_eq!(results[1].street_rank_results[2].num_iterations, num_it);
-        assert_eq!(3, results[1].player_index);
-
-        assert_equity(
-            100.0 * results[1].street_rank_results[2].win_eq / num_it as f64,
-            78.08,
-            tolerance,
-        );
-    }
-
-    #[test]
-    fn test_3way_with_ranges() {
-        let mut analyzer = super::flop_analyzer::new();
-        analyzer.reset();
-
-        analyzer.set_player_state(0, PlayerPreFlopState::UseHoleCards as u8);
-        analyzer.set_player_state(2, PlayerPreFlopState::UseRange as u8);
-        analyzer.set_player_state(3, PlayerPreFlopState::UseHoleCards as u8);
-
-        analyzer
-            .set_player_cards(0, card_u8s_from_string("8d 7s").as_slice())
-            .unwrap();
-
-        analyzer
-            .set_player_cards(3, card_u8s_from_string("Qd 5c").as_slice())
-            .unwrap();
-
-        analyzer.set_player_range(
-            2,
-            "22+, A2s+, K2s+, Q2s+, J6s+, 94s, A2o+, K7o+, QJo, J7o, T4o",
-        );
-
-        analyzer
-            .set_board_cards(card_u8s_from_string("Qs Ts 7c").as_slice())
-            .unwrap();
-
-        let num_it = 4_000;
-
-        let tolerance = 0.5;
-        //let tolerance = 0.1;
-
-        let f_results = analyzer.build_results();
-        let f_results = analyzer.simulate_flop(num_it, f_results).unwrap();
-
-        let results = &f_results.flop_results;
-
-        assert_eq!(results[0].street_rank_results[2].num_iterations, num_it);
-        assert_eq!(results[0].player_index, 0);
-
-        assert_equity(
-            100.0 * results[0].street_rank_results[2].win_eq
-                / results[0].street_rank_results[2].num_iterations as f64,
-            21.03,
-            tolerance,
-        );
-        assert_equity(
-            100.0 * results[0].street_rank_results[2].tie_eq
-                / results[0].street_rank_results[2].num_iterations as f64,
-            0.12,
-            0.05,
-        );
-
-        assert_eq!(results[2].street_rank_results[2].num_iterations, num_it);
-        assert_eq!(results[2].player_index, 3);
-
-        // assert_equity(
-        //     100.0 * results[3].eq_not_folded / not_folded as f64,
-        //     50.93 + 0.82,
-        //     0.7,
-        // );
-        assert_equity(
-            100.0 * results[2].street_rank_results[2].win_eq / num_it as f64,
-            50.93,
-            tolerance,
-        );
-        assert_equity(
-            100.0 * results[2].street_rank_results[2].tie_eq / num_it as f64,
-            0.82,
-            tolerance,
-        );
-
-        assert_eq!(results[1].street_rank_results[2].num_iterations, num_it);
-        assert_eq!(results[1].player_index, 2);
-        //let not_folded = results[3].num_iterations;
-
-        assert_equity(
-            100.0 * results[1].street_rank_results[2].win_eq / num_it as f64,
-            26.14,
-            tolerance,
-        );
-        assert_equity(
-            100.0 * results[1].street_rank_results[2].tie_eq / num_it as f64,
-            0.95,
-            tolerance,
-        );
-    }
-
-    #[test]
-    fn test_villian_draws() {
-        let mut analyzer = super::flop_analyzer::new();
-        analyzer.reset();
-
-        analyzer.set_player_state(0, PlayerPreFlopState::UseHoleCards as u8);
-        analyzer.set_player_state(3, PlayerPreFlopState::UseHoleCards as u8);
-        analyzer.set_player_state(4, PlayerPreFlopState::UseHoleCards as u8);
-        analyzer.set_player_state(2, PlayerPreFlopState::UseHoleCards as u8);
-
-        analyzer
-            .set_player_cards(0, card_u8s_from_string("Td 8s").as_slice())
-            .unwrap();
-
-        analyzer
-            .set_player_cards(3, card_u8s_from_string("Ad Kc").as_slice())
-            .unwrap();
-        analyzer
-            .set_player_cards(4, card_u8s_from_string("5s 5c").as_slice())
-            .unwrap();
-        analyzer
-            .set_player_cards(2, card_u8s_from_string("Qd 7d").as_slice())
-            .unwrap();
-
-        analyzer
-            .set_board_cards(card_u8s_from_string("9s 8c Ah 5h 6h").as_slice())
-            .unwrap();
-
-        let num_it = 1;
-
-        let results = analyzer.build_results();
-        let results = analyzer.simulate_flop(num_it, results).unwrap();
-
-        let v_r = &results.all_villians;
-        assert_eq!(
-            1,
-            v_r.street_rank_results[0].rank_family_count[Rank::OnePair(0).get_family_index()]
-        );
-        assert_eq!(
-            1u32,
-            v_r.street_rank_results[0].rank_family_count.iter().sum()
-        );
-        assert_eq!(0, v_r.street_draws[0].gut_shot);
-        assert_eq!(0, v_r.street_draws[0].two_overcards);
-        assert_eq!(0, v_r.street_draws[0].one_overcard);
-        assert_eq!(
-            1.0,
-            v_r.street_rank_results[0].win_eq / v_r.street_rank_results[0].num_iterations as f64
-        );
-
-        //Turn villian picks up gut shot
-        assert_eq!(
-            1,
-            v_r.street_rank_results[1].rank_family_count[Rank::ThreeOfAKind(0).get_family_index()]
-        );
-        assert_eq!(
-            1u32,
-            v_r.street_rank_results[1].rank_family_count.iter().sum()
-        );
-        assert_eq!(1, v_r.street_draws[1].gut_shot);
-        assert_eq!(0, v_r.street_draws[1].two_overcards);
-        assert_eq!(0, v_r.street_draws[1].one_overcard);
-
-        assert_eq!(
-            0,
-            v_r.street_rank_results[2].rank_family_count[Rank::OnePair(0).get_family_index()]
-        );
-        assert_eq!(
-            1,
-            v_r.street_rank_results[2].rank_family_count[Rank::Straight(0).get_family_index()]
-        );
-        assert_eq!(
-            1u32,
-            v_r.street_rank_results[2].rank_family_count.iter().sum()
-        );
-        assert_eq!(2, v_r.street_draws.len());
-    }
-
-    #[test]
-    fn test_villian_overcards() {
-        let mut analyzer = super::flop_analyzer::new();
-        analyzer.reset();
-
-        analyzer.set_player_state(0, PlayerPreFlopState::UseHoleCards as u8);
-        analyzer.set_player_state(3, PlayerPreFlopState::UseHoleCards as u8);
-        analyzer.set_player_state(4, PlayerPreFlopState::UseHoleCards as u8);
-        analyzer.set_player_state(2, PlayerPreFlopState::UseHoleCards as u8);
-
-        analyzer
-            .set_player_cards(0, card_u8s_from_string("Tc 8s").as_slice())
-            .unwrap();
-
-        analyzer
-            .set_player_cards(3, card_u8s_from_string("Ad Jc").as_slice())
-            .unwrap();
-        analyzer
-            .set_player_cards(4, card_u8s_from_string("Ks Qc").as_slice())
-            .unwrap();
-        analyzer
-            .set_player_cards(2, card_u8s_from_string("Jd Td").as_slice())
-            .unwrap();
-
-        analyzer
-            .set_board_cards(card_u8s_from_string("2s 4c 7h Qh Ah").as_slice())
-            .unwrap();
-
-        let num_it = 1;
-
-        let results = analyzer.build_results();
-        let results = analyzer.simulate_flop(num_it, results).unwrap();
-
-        let v_r = &results.all_villians;
-        assert_eq!(
-            1,
-            v_r.street_rank_results[0].rank_family_count[Rank::HighCard(0).get_family_index()]
-        );
-        assert_eq!(
-            1,
-            v_r.street_rank_results[0]
-                .rank_family_count
-                .iter()
-                .sum::<u32>()
-        );
-        assert_eq!(1, v_r.street_draws[0].two_overcards);
-        assert_eq!(0, v_r.street_draws[0].one_overcard);
-
-        assert_eq!(
-            1,
-            v_r.street_rank_results[1].rank_family_count[Rank::OnePair(0).get_family_index()]
-        );
-        assert_eq!(
-            1,
-            v_r.street_rank_results[1]
-                .rank_family_count
-                .iter()
-                .sum::<u32>()
-        );
-        assert_eq!(0, v_r.street_draws[1].two_overcards);
-        assert_eq!(1, v_r.street_draws[1].one_overcard);
-
-        assert_eq!(
-            1,
-            v_r.street_rank_results[2].rank_family_count[Rank::OnePair(0).get_family_index()]
-        );
-        assert_eq!(
-            1,
-            v_r.street_rank_results[2]
-                .rank_family_count
-                .iter()
-                .sum::<u32>()
-        );
-        assert_eq!(2, v_r.street_draws.len());
     }
 }
